Refactor helpers.
-- helpers.rb
- Removed all the connection-specific helpers to their own file:
  connection_helpers.rb. The reason for doing this is to keep the
  different helper-files small.

-- jdbc_(ora | pg)_connection.rb and (oci | pg)_connection.rb
- Changed the reference of helpers.rb to connection_helpers.rb.
