+++ lib/plsql/jdbc_pg_connection.rb
@@ -24,7 +24,7 @@ rescue LoadError, NameError
   raise LoadError
 end
 
-require "plsql/helpers"
+require "plsql/connection_helpers"
 
 module PLSQL
   class JDBCPGConnection < Connection #:nodoc:
